Refactor ValidationValue to allow rtol/atol in MetaValidator
